Clean up people filter, better seeding.
